Expand the map format

+++ src/fheroes2/gui/ui_language.h
@@ -25,7 +25,7 @@
 
 namespace fheroes2
 {
-    enum class SupportedLanguage : int
+    enum class SupportedLanguage : uint8_t
     {
         English = 0, // default language for all versions of the game.
         French, // GOG version
